Atualização back-end
Arquivos model, repository e service minimamente configurados. Adicionados arquivo config para configurações de segurança, enums para postos de graduação e roles e uma exception.

+++ codigo/project/src/main/java/com/WorkFlowManager/project/dto/OrganizacaoDTO.java
@@ -2,9 +2,8 @@ package com.WorkFlowManager.project.dto;
 
 public record OrganizacaoDTO
 (
-    int id,
-    int[] idSubOrganizacoes,
-
+    Long id,
+    Long[] idSubOrganizacoes,
     String nome,
     String abreviacao
 ) {}
